+++ src/main/java/org/webbitserver/wrapper/HttpRequestWrapper.java
@@ -1,9 +1,11 @@
 package org.webbitserver.wrapper;
 
 import org.webbitserver.HttpRequest;
+import org.webbitserver.helpers.QueryParameters;
 
 import java.net.HttpCookie;
 import java.net.SocketAddress;
+import java.net.URI;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -40,6 +42,12 @@ public class HttpRequestWrapper implements HttpRequest {
     }
 
     @Override
+    public HttpRequestWrapper uri(String uri) {
+        request.uri(uri);
+        return this;
+    }
+
+    @Override
     public String header(String name) {
         return request.header(name);
     }
@@ -57,6 +65,16 @@ public class HttpRequestWrapper implements HttpRequest {
     @Override
     public HttpCookie cookie(String name) {
         return request.cookie(name);
+    }
+
+    @Override
+    public String queryParam(String key) {
+        return request.queryParam(key);
+    }
+
+    @Override
+    public List<String> queryParams(String key) {
+        return request.queryParams(key);
     }
 
     @Override